restructure checks
