+++ doubleml/double_ml_irm.py
@@ -10,8 +10,8 @@ from .double_ml_blp import DoubleMLBLP
 from .double_ml_data import DoubleMLData
 from .double_ml_score_mixins import LinearScoreMixin
 
-from ._utils import _dml_cv_predict, _get_cond_smpls, _dml_tune, _check_finite_predictions, _check_is_propensity, \
-    _trimm, _normalize_ipw, _check_score, _check_trimming
+from ._utils import _dml_cv_predict, _get_cond_smpls, _dml_tune, _trimm, _normalize_ipw
+from ._utils_checks import _check_score, _check_trimming, _check_finite_predictions, _check_is_propensity
 
 
 class DoubleMLIRM(LinearScoreMixin, DoubleML):
