+++ doubleml/tests/test_nonlinear_score_mixin.py
@@ -8,7 +8,8 @@ from sklearn.utils import check_X_y
 
 import doubleml as dml
 from doubleml.double_ml import DoubleML
-from doubleml._utils import _dml_cv_predict, _check_finite_predictions
+from doubleml._utils import _dml_cv_predict
+from doubleml._utils_checks import _check_finite_predictions
 from doubleml.double_ml_score_mixins import NonLinearScoreMixin
